add data corruption telemetry to remote channel context
