+++ packages/loader/container-utils/package.json
@@ -51,6 +51,7 @@
     "temp-directory": "nyc/.nyc_output"
   },
   "dependencies": {
+    "@fluidframework/common-definitions": "^0.19.1",
     "@fluidframework/common-utils": "^0.25.0",
     "@fluidframework/container-definitions": "^0.30.0",
     "@fluidframework/telemetry-utils": "^0.30.0",
